feat(llm): add TokenUsage tracking to LLMResponse
Add TokenUsage struct to track prompt_tokens, completion_tokens, and
total_tokens from LLM API responses. This enables monitoring token
consumption across different providers.

- Add TokenUsage struct with new() constructor that auto-calculates total
- Add usage: Option<TokenUsage> field to LLMResponse
- Extract token usage from OpenAI response.usage
- Extract token usage from Ollama final_data (prompt_eval_count, eval_count)
- Add usage: None for LlamaCpp (not exposed by llama-cpp-2 crate)
- Update mock clients in tests

+++ src/llm/openai.rs
@@ -356,7 +356,7 @@ impl LLMClient for OpenAIClient {
         // Extract token usage if available
         let usage = response
             .usage
-            .map(|u| TokenUsage::new(u.prompt_tokens, u.completion_tokens));
+            .map(|u| TokenUsage::new(u.prompt_tokens as u32, u.completion_tokens as u32));
 
         Ok(LLMResponse {
             content,
